apps/apis/v1/store/serializers.py
from rest_framework import serializers

from apps.store.models import Document


class DocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Document
        fields = ("id", "owner")
