requirements.txt
Django==4.0.1
djangorestframework==3.13.1
python-decouple==3.3

pre-commit==2.17.0

factory-boy==3.2.1
pytest==6.2.5
pytest-sugar==0.9.4


